+++ include/ArduinoJson/Internals/TextWriter.hpp
@@ -8,8 +8,7 @@
 #pragma once
 
 #include "../Polyfills/attributes.hpp"
-#include "../Polyfills/isInfinity.hpp"
-#include "../Polyfills/isNaN.hpp"
+#include "../Polyfills/math.hpp"
 #include "../Polyfills/normalize.hpp"
 #include "Encoding.hpp"
 #include "JsonFloat.hpp"
@@ -78,12 +77,6 @@ class TextWriter {
       powersOf10 = 0;
     }
 
-    // Round correctly so that print(1.999, 2) prints as "2.00"
-    JsonFloat rounding = 0.5;
-    for (int i = 0; i < digits; ++i) rounding /= 10.0;
-
-    value += rounding;
-
     // Extract the integer part of the value and print it
     JsonUInt int_part = static_cast<JsonUInt>(value);
     JsonFloat remainder = value - static_cast<JsonFloat>(int_part);
@@ -96,10 +89,16 @@ class TextWriter {
 
     // Extract digits from the remainder one at a time
     while (digits-- > 0) {
+      // Extract digit
       remainder *= 10.0;
-      JsonUInt toPrint = JsonUInt(remainder);
-      writeInteger(JsonUInt(remainder));
-      remainder -= static_cast<JsonFloat>(toPrint);
+      char currentDigit = char(remainder);
+      remainder -= static_cast<JsonFloat>(currentDigit);
+
+      // Round up last digit (so that print(1.999, 2) prints as "2.00")
+      if (digits == 0 && remainder >= 0.5) currentDigit++;
+
+      // Print
+      writeRaw(char('0' + currentDigit));
     }
 
     if (powersOf10 < 0) {